+++ packages/core/src/__tests__/test-generator.ts
@@ -276,6 +276,10 @@ const IMPORT_TEST: Tests = {
 const ANGULAR_TESTS: Tests = {
   nativeAttributes: getRawFile('./data/angular/native-attributes.raw.tsx'),
   outputEventBinding: getRawFile('./data/angular/output-event-bindings.raw.tsx'),
+  dynamicComponent: getRawFile('./data/angular/dynamic-component.raw.tsx'),
+  dynamicComponentWithEventArg: getRawFile(
+    './data/angular/dynamic-component-with-event-args.raw.tsx',
+  ),
 };
 
 const CONTEXT_TEST: Tests = {