+++ packages/docs/src/routes/playground/index.tsx
@@ -1,111 +1,38 @@
-import { MitosisComponent } from '@builder.io/mitosis';
 import { $, component$, useSignal, useVisibleTask$ } from '@builder.io/qwik';
-import { routeLoader$, server$, useLocation } from '@builder.io/qwik-city';
+import { DocumentHead, routeLoader$, useLocation, useNavigate } from '@builder.io/qwik-city';
+import lzString from 'lz-string';
 import { ContentLoaderCode } from 'qwik-content-loader';
 import { CodeEditor } from '~/components/code-editor';
 import Select from '~/components/select';
-
-export type OutputFramework = 'react' | 'svelte' | 'vue' | 'qwik' | 'angular' | 'mitosis' | 'json';
-const outputs: OutputFramework[] = ['react', 'svelte', 'vue', 'qwik', 'angular', 'mitosis', 'json'];
-
-export type InputSyntax = 'jsx' | 'svelte';
-const inputs: InputSyntax[] = ['jsx', 'svelte'];
+import {
+  InputSyntax,
+  OutputFramework,
+  compile,
+  defaultCode,
+  inputs,
+  languageByFramework,
+  outputs,
+} from '~/services/compile';
 
 const defaultTopTab: OutputFramework = 'vue';
 const defaultBottomTab: OutputFramework = 'angular';
 const defaultInputTab = 'jsx';
 
-const languageByFramework: Record<OutputFramework, string> = {
-  react: 'typescript',
-  svelte: 'html',
-  vue: 'html',
-  qwik: 'typescript',
-  angular: 'typescript',
-  mitosis: 'typescript',
-  json: 'json',
+const decodeCode = (url: URL) => {
+  const code = url.searchParams.get('code');
+  return code ? lzString.decompressFromBase64(code) : defaultCode;
 };
-
-const getOutputGenerator = async ({ output }: { output: OutputFramework }) => {
-  const {
-    parseJsx,
-    componentToSvelte,
-    componentToVue,
-    componentToReact,
-    componentToQwik,
-    componentToAngular,
-    componentToMitosis,
-    parseSvelte,
-  } = await import('@builder.io/mitosis');
-
-  const options = {};
-
-  switch (output) {
-    case 'qwik':
-      return componentToQwik(options);
-    case 'react':
-      return componentToReact(options);
-    case 'angular':
-      return componentToAngular(options);
-    case 'svelte':
-      return componentToSvelte(options);
-    case 'mitosis':
-      return componentToMitosis();
-    case 'json':
-      return ({ component }: { component: MitosisComponent }) => JSON.stringify(component, null, 2);
-    case 'vue':
-      return componentToVue({ api: 'composition' });
-    default:
-      throw new Error('unexpected Output ' + output);
-  }
-};
-
-export const compile = server$(
-  async (code: string, output: OutputFramework, inputSyntax: InputSyntax) => {
-    const {
-      parseJsx,
-      componentToSvelte,
-      componentToVue,
-      componentToReact,
-      componentToQwik,
-      componentToAngular,
-      componentToMitosis,
-      parseSvelte,
-    } = await import('@builder.io/mitosis');
-    const parsed = inputSyntax === 'svelte' ? await parseSvelte(code) : parseJsx(code);
-
-    const outputGenerator = await getOutputGenerator({ output });
-
-    const outputCode = outputGenerator({ component: parsed });
-
-    return outputCode;
-  },
-);
-
-const defaultCode = `
-import { useState } from "@builder.io/mitosis";
-
-export default function MyComponent(props) {
-  const [name, setName] = useState("Steve");
-
-  return (
-    <div>
-      <input
-        css={{
-          color: "red",
-        }}
-        value={name}
-        onChange={(event) => setName(event.target.value)}
-      />
-      Hello! I can run in React, Vue, Solid, or Liquid!
-    </div>
-  );
-}
-`.trim();
 
 const useOutput1 = routeLoader$(async (requestEvent) => {
-  const code = (requestEvent.url.searchParams.get('code') as string) || defaultCode;
-  const outputTab = requestEvent.url.searchParams.get('outputTab') as OutputFramework;
-  const inputTab = requestEvent.url.searchParams.get('inputTab') as InputSyntax;
+  const code = decodeCode(requestEvent.url);
+  let outputTab = requestEvent.url.searchParams.get('outputTab') as OutputFramework;
+  if (!outputs.includes(outputTab)) {
+    outputTab = defaultTopTab;
+  }
+  let inputTab = requestEvent.url.searchParams.get('inputTab') as InputSyntax;
+  if (!inputs.includes(inputTab)) {
+    inputTab = defaultInputTab;
+  }
 
   const output = await compile(
     code || defaultCode,
@@ -119,12 +46,12 @@ const useOutput1 = routeLoader$(async (requestEvent) => {
 });
 
 const useOutput2 = routeLoader$(async (requestEvent) => {
-  const code = (requestEvent.url.searchParams.get('code') as string) || defaultCode;
+  const code = decodeCode(requestEvent.url);
   const outputTab = requestEvent.url.searchParams.get('outputTab') as OutputFramework;
   const inputTab = requestEvent.url.searchParams.get('inputTab') as InputSyntax;
 
   const output = await compile(
-    code || defaultCode,
+    code,
     outputTab || defaultBottomTab,
     inputTab || defaultInputTab,
   ).catch((err) => {
@@ -135,14 +62,15 @@ const useOutput2 = routeLoader$(async (requestEvent) => {
 });
 
 export default component$(() => {
+  const nav = useNavigate();
   const location = useLocation();
-  const codeFromQueryParam = location.url.searchParams.get('code') as string;
+  const initialCode = decodeCode(location.url);
   const outputTab = location.url.searchParams.get('outputTab') as OutputFramework;
   const inputTab = location.url.searchParams.get('inputTab') as InputSyntax;
   const loaderOutput1 = useOutput1().value;
   const loaderOutput2 = useOutput2().value;
 
-  const code = useSignal(codeFromQueryParam || defaultCode);
+  const code = useSignal(initialCode || defaultCode);
   const inputSyntax = useSignal<InputSyntax>(inputTab || defaultInputTab);
   const output = useSignal(loaderOutput1 || '');
   const outputOneFramework = useSignal<OutputFramework>(outputTab || defaultTopTab);
@@ -158,8 +86,28 @@ export default component$(() => {
     visible.value = true;
   });
 
+  const updateUrl = $(() => {
+    if (code.value === defaultCode || !code.value.trim()) {
+      if (location.url.searchParams.has('code')) {
+        location.url.searchParams.delete('code');
+        nav(location.url.toString(), {
+          replaceState: true,
+        });
+      }
+      return;
+    }
+
+    const newURL = new URL(location.url);
+    newURL.searchParams.set('code', lzString.compressToBase64(code.value));
+
+    nav(newURL.toString(), {
+      replaceState: true,
+    });
+  });
+
   const throttledCompileOne = $(
     async (code: string, outputFramework: OutputFramework, inputSyntax: InputSyntax) => {
+      updateUrl();
       if (throttleTimeout1.value) {
         clearTimeout(throttleTimeout1.value);
       }
@@ -167,6 +115,7 @@ export default component$(() => {
         throttleTimeout1.value = setTimeout(async () => {
           isThrottling.value = true;
           output.value = await compile(code, outputFramework, inputSyntax);
+          updateUrl();
           isThrottling.value = false;
         }, 100) as any;
         return;
@@ -229,15 +178,16 @@ export default component$(() => {
   });
 
   return (
-    <div class="relative flex gap-4 grow items-stretch max-md:flex-col">
-      <div class="w-full flex flex-col max-md:h-[50vh]">
-        <div class="flex items-center gap-2 mx-4 my-4 mb-0 min-h-[50px]">
-          <h3 class="ml-4 text-lg">Input</h3>
+    <div class="relative flex gap-4 max-md:gap-0 grow items-stretch max-md:flex-col bg-primary-dark overflow-x-hidden">
+      <style>{`body { overflow: hidden !important; }`}</style>
+      <div class="w-full flex flex-col max-md:h-[calc(55dvh-35px)]">
+        <div class="flex items-center gap-2 mx-4 my-2 mb-4 max-md:m-1.5 min-h-[50px] max-md:min-h-[40px]">
+          <h3 class="ml-4 max-md:ml-2 text-lg max-md:text-base">Input</h3>
           {visible.value && (
             // Workaround weird bug where this doesn't render correctly
             // server side
             <Select
-              class="ml-auto"
+              class="ml-auto max-md:scale-[0.85] -my-2 max-md:-mr-1.5"
               value={inputSyntax.value}
               onChange$={(syntax: any) => {
                 compile(
@@ -255,7 +205,10 @@ export default component$(() => {
         </div>
 
         <div class="w-full grow relative">
-          <ContentLoaderCode width={400} class="ml-16 mt-3 opacity-10 origin-top-left" />
+          <ContentLoaderCode
+            width={400}
+            class="ml-16 mt-3 opacity-10 origin-top-left max-md:scale-75 max-md:ml-4"
+          />
 
           {visible.value && (
             <CodeEditor
@@ -270,22 +223,25 @@ export default component$(() => {
           )}
         </div>
       </div>
-      <div class="flex gap-4 flex-col w-full h-[90vh] max-md:h-[50vh]">
-        <div class="flex items-center gap-2 mx-4 my-4 mb-0 min-h-[50px]">
-          <h3 class="ml-4 text-lg">Output</h3>
+      <div class="flex gap-4 max-md:gap-0 flex-col w-full h-[90vh] max-md:!h-[calc(45dvh-35px)] border-l border-primary border-opacity-50 max-md:border-l-0 max-md:border-t">
+        <div class="flex items-center gap-2 mx-4 max-md:m-1.5 my-2 mb-0 min-h-[50px] max-md:min-h-[40px]">
+          <h3 class="ml-4 max-md:ml-2 text-lg max-md:text-base">Output</h3>
           {visible.value && (
             // Workaround weird bug where this doesn't render correctly
             // server side
             <Select
-              class="ml-auto mr-2"
+              class="ml-auto mr-2 max-md:scale-[0.85] mx-md:-my-2 max-md:-mr-1.5"
               value={outputOneFramework.value}
               onChange$={(framework: any) => (outputOneFramework.value = framework)}
               options={outputs}
             />
           )}
         </div>
-        <div class="h-[50%] relative">
-          <ContentLoaderCode width={400} class="ml-16 mt-3 opacity-10 origin-top-left" />
+        <div class="h-[50%] max-md:h-auto grow relative">
+          <ContentLoaderCode
+            width={400}
+            class="ml-16 mt-3 opacity-10 origin-top-left max-md:scale-75 max-md:ml-4"
+          />
           {visible.value && (
             <CodeEditor
               language={languageByFramework[outputOneFramework.value]}
@@ -295,12 +251,12 @@ export default component$(() => {
             />
           )}
         </div>
-        <div class="min-h-[50px] max-md:hidden flex items-center">
+        <div class="min-h-[50px] max-md:min-h-[40px] max-md:hidden flex items-center border-primary border-opacity-50 border-t -mt-4 pt-4">
           {visible.value && (
             // Workaround weird bug where this doesn't render correctly
             // server side
             <Select
-              class="ml-auto mr-2"
+              class="ml-auto mr-2 md:mr-6"
               value={outputTwoFramework.value}
               onChange$={(framework: any) => (outputTwoFramework.value = framework)}
               options={outputs}
@@ -309,7 +265,10 @@ export default component$(() => {
         </div>
 
         <div class="h-[50%] relative max-md:hidden">
-          <ContentLoaderCode width={400} class="ml-16 mt-3 opacity-10 origin-top-left" />
+          <ContentLoaderCode
+            width={400}
+            class="ml-16 mt-3 opacity-10 origin-top-left max-md:scale-75 max-md:ml-4"
+          />
           {visible.value && (
             <CodeEditor
               language={languageByFramework[outputTwoFramework.value]}
@@ -323,3 +282,13 @@ export default component$(() => {
     </div>
   );
 });
+
+export const head: DocumentHead = {
+  title: 'Mitosis Playground',
+  meta: [
+    {
+      name: 'description',
+      content: 'Write components once, run everywhere.',
+    },
+  ],
+};