+++ packages/docs/vite.config.ts
@@ -2,20 +2,21 @@
  * This is the base config for vite.
  * When building, the adapter config is used which loads this file and extends it.
  */
+import { partytownVite } from '@builder.io/partytown/utils';
 import { qwikCity } from '@builder.io/qwik-city/vite';
+import { qwikInsights } from '@builder.io/qwik-labs-canary/vite';
 import { qwikVite } from '@builder.io/qwik/optimizer';
+import { join } from 'path';
 import { defineConfig, type UserConfig } from 'vite';
 import { nodePolyfills } from 'vite-plugin-node-polyfills';
 import { replaceCodePlugin } from 'vite-plugin-replace';
 import tsconfigPaths from 'vite-tsconfig-paths';
 import pkg from './package.json';
-
 const { dependencies = {}, devDependencies = {} } = pkg as any as {
   dependencies: Record<string, string>;
   devDependencies: Record<string, string>;
   [key: string]: unknown;
 };
-
 /**
  * Note that Vite normally starts from `index.html` but the qwikCity plugin makes start at `src/entry.ssr.tsx` instead.
  */
@@ -29,6 +30,9 @@ export default defineConfig(({ command, mode }): UserConfig => {
       qwikVite(),
       tsconfigPaths({ root: './' }),
       nodePolyfills(),
+      qwikInsights({
+        publicApiKey: '22gsbhtjcyv',
+      }),
       replaceCodePlugin({
         replacements: [
           {
@@ -37,6 +41,7 @@ export default defineConfig(({ command, mode }): UserConfig => {
           },
         ],
       }),
+      partytownVite({ dest: join(__dirname, 'dist', '~partytown') }),
     ],
     // This tells Vite which dependencies to pre-build in dev mode.
     optimizeDeps: {